actualización pern: inicio react, componentes, rutas y páginas.

+++ Javascript/pern-stack/backend/src/controllers/constants.js
@@ -5,7 +5,7 @@ const CREATE_TASK =
   "INSERT INTO tareas(titulo, descripcion, user_id) VALUES($1, $2, $3) RETURNING *;";
 const UPDATE_TASK =
   "UPDATE tareas SET titulo = $2, descripcion = $3, actualizado = $4, actualizado_el = $5 WHERE tarea_id = $1 RETURNING *;";
-const DELETE_TASK = "DELETE FROM tareas WHERE tarea_id = $1 RETURNING *;";
+const DELETE_TASK = "DELETE FROM tareas WHERE tarea_id = $1 AND user_id = $2 RETURNING *;";
 const GET_ALL_USERS = "SELECT * FROM pern_user;";
 const CREATE_USER =
   "INSERT INTO pern_user (user_name, user_email, user_password, user_avatar) VALUES($1, $2, $3, $4) RETURNING *;";
